fixed csv parsing. added csv string parsing
