+++ server/app/common/services/file.services/file.service.js
@@ -2,16 +2,17 @@ const CsvService = require('./csv.service');
 const XlsxService = require('./xlsx.service');
 
 module.exports = {
-	fileService: file =>
+	fileService: (file, contents) =>
 		new Promise((resolve, reject) => {
 			const pattern = {
 				csv: /.+(\.csv)$/,
 				xlsx: /.+(\.xlsx)$/
 			};
-			if (!file) {
+			if (contents !== '' && typeof contents === 'string') {
+				resolve(CsvService.processString(contents));
+			} else if (!file) {
 				throw new Error('No file were uploaded');
-			}
-			if (pattern.csv.test(file.fileKey.name)) {
+			} else if (pattern.csv.test(file.fileKey.name)) {
 				resolve(CsvService.processFile(file.fileKey));
 			} else if (pattern.xlsx.test(file.fileKey.name)) {
 				resolve(XlsxService.processFile(file.fileKey));
