+++ server/app/common/services/file.services/fs.service.js
@@ -1,22 +1,25 @@
 const fs = require('fs');
+const uuidv4 = require('uuid/v4');
 
 class FsService {
 	save(file) {
+		let path = `../server/app/fileStorage/${uuidv4()}${file.name}`;
 		return new Promise((resolve, reject) => {
 			// todo: rename file if exists otherwise it will be overwritten
-			if (fs.existsSync(`../server/app/fileStorage/${file.name}`)) {
-				return null;
+			while (fs.existsSync(path)) {
+				path = `../server/app/fileStorage/${uuidv4()}${file.name}`;
 			}
-			const path = `../server/app/fileStorage/${file.name}`;
 			const stream = fs.createWriteStream(path);
 			return stream.once('open', () => {
 				stream.write(file.data);
 				stream.end(() => {
-					resolve(`../server/app/fileStorage/${file.name}`);
+					resolve(path);
 				});
 				stream.on('error', () => {
-					this.deleteFile(path);
-					reject(new Error('error write stream'));
+					this.deleteFile(path).then(() => {
+						reject(new Error("File wasn't processed"));
+					});
+					reject(new Error('Error write stream'));
 				});
 			});
 		});
