+++ server/app/fileStorage/171ecb26-9360-4d42-b69d-0d2f114226bcbad.csv
@@ -1,0 +1,10 @@
+Product name,Price,Quantity, ,FF
+Mollis consequat,9,2,
+Tvoluptatem,10.32,1,
+Tvoluptatem,10.32,1,
+Broken,,1,
+Broken1,1,
+Tvoluptatem,10.32,1,
+Scelerisque lacinia,,1,
+Consectetur adipiscing,28.72,10,
+Condimentum aliquet,13.9,1,wtf,fsd,gew,g,wegwe,htr
