+++ server/package.json
@@ -32,6 +32,8 @@
 		"pg-hstore": "^2.3.2",
 		"pretty-quick": "^1.6.0",
 		"sequelize": "^4.38.0",
+		"tempfile": "^2.0.0",
+		"uuid": "^3.3.2",
 		"xlsx": "^0.13.3"
 	},
 	"devDependencies": {
